% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.SDAM}
\alias{predict.SDAM}
\title{Predictions for SDAM}
\usage{
\method{predict}{SDAM}(object, Xnew, ...)
}
\arguments{
\item{object}{Fitted object of class \code{SDAM}.}

\item{Xnew}{Matrix of new test data at which to evaluate the fitted function.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A vector of predictions for the new data.
}
\description{
Predicts the response for new data using a fitted SDAM.
}
\examples{
set.seed(1)
X <- matrix(rnorm(20 * 15), ncol = 15)
Y <- sin(X[, 1]) -  X[, 2] + rnorm(20)
model <- SDAM(X, Y, Q_type = "trim", trim_quantile = 0.5, cv_k = 5)
predict(object = model, Xnew = X)
}
\seealso{
\code{\link{SDAM}}
}
\author{
Cyrill Scheidegger
}
