% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_individual_fj}
\alias{predict_individual_fj}
\title{Predictions of individual component functions for SDAM}
\usage{
predict_individual_fj(object, Xjnew, j)
}
\arguments{
\item{object}{Fitted object of class \code{SDAM}.}

\item{Xjnew}{Vector of new test data at which to evaluate fj.}

\item{j}{Which component to evaluate.}
}
\value{
A vector of predictions for fj evaluated at Xjnew.
}
\description{
Predicts the contribution of an individual component j using a fitted SDAM.
}
\examples{
set.seed(1)
X <- matrix(rnorm(20 * 15), ncol = 15)
Y <- sin(X[, 1]) -  X[, 2] + rnorm(20)
model <- SDAM(X, Y, Q_type = "trim", trim_quantile = 0.5, cv_k = 5)
predict_individual_fj(object = model, Xjnew = X[, 1], j = 1)
}
\seealso{
\code{\link{SDAM}}
}
\author{
Cyrill Scheidegger
}
