% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partDependence.R
\name{print.partDependence}
\alias{print.partDependence}
\title{Print partDependence}
\usage{
\method{print}{partDependence}(x, ...)
}
\arguments{
\item{x}{Fitted object of class \code{partDependence}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of the partDependence.
}
\seealso{
\code{\link{partDependence}}, \code{\link{plot.partDependence}}
}
\author{
Markus Ulmer
}
