% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SDAM}
\alias{print.SDAM}
\title{Print SDAM}
\usage{
\method{print}{SDAM}(x, ...)
}
\arguments{
\item{x}{Fitted object of class \code{SDAM}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print number of covariates and number of active covariates for SDAM object.
}
\seealso{
\code{\link{SDAM}}
}
\author{
Cyrill Scheidegger
}
