% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SDForest}
\alias{print.SDForest}
\title{Print SDForest}
\usage{
\method{print}{SDForest}(x, ...)
}
\arguments{
\item{x}{Fitted object of class \code{SDForest}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
Print contents of the SDForest.
}
\examples{

set.seed(1)
n <- 50
X <- matrix(rnorm(n * 5), nrow = n)
y <- sign(X[, 1]) * 3 + rnorm(n)
model <- SDForest(x = X, y = y, Q_type = 'no_deconfounding', nTree = 5, cp = 0.5)
print(model)
}
\seealso{
\code{\link{SDForest}}
}
\author{
Markus Ulmer
}
