% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partDependence.R
\name{print.partDependence}
\alias{print.partDependence}
\title{Print partDependence}
\usage{
\method{print}{partDependence}(x, ...)
}
\arguments{
\item{x}{Fitted object of class \code{partDependence}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
Print contents of the partDependence.
}
\examples{
set.seed(1)
x <- rnorm(10)
y <- sign(x) * 3 + rnorm(10)
model <- SDTree(x = x, y = y, Q_type = 'no_deconfounding', cp = 0.5)
pd <- partDependence(model, 1, X = x)
print(pd)
}
\seealso{
\code{\link{partDependence}}, \code{\link{plot.partDependence}}
}
\author{
Markus Ulmer
}
